deleted a bunch of old files and moved new files around
